refactor(dashboard): extract release-provider-matrix notice components (#14397)
## Summary

Repo-quality loop, iteration 7 (no Linear issue — ad-hoc loop; Linear still at free-issue limit). Behavior-preserving split of `ReleaseProviderMatrix.tsx` (**1388 → 1157 lines**).

- The 5 clean, prop-driven notice/dialog presentation components — `ReleaseImportProgressNotice`, `ReleaseAppleMusicSyncNotice`, `ReleaseSmartLinkNotices`, `ConnectedReleaseEmptyState`, `ReleasePlanDialog` — move verbatim to a new sibling `ReleaseProviderMatrixNotices.tsx` (same `release-provider-matrix/` dir — eslint-boundaries safe). The new file re-declares its own `lazy()` imports of the banner sibling modules and the `SMART_LINK_SOFT_CAP` constant; biome pruned the now-unused lazy defs from the main file.
- `AppleMusicPaletteDialog` and `DeleteReleaseConfirmation` deliberately **stay** in the main file — they reference the shared `AppleMusicArtist` type / `isDistributedRelease` helper, so moving them would create a circular import. Extracting only the 5 clean components avoids any shared-symbol move.
- Main file imports the 5 components back. Bodies verbatim, no logic changes.

## Verification

- `pnpm biome check` → **clean ✓**
- `pnpm --filter @jovie/web run typecheck` → **pass ✓** (0 errors)
- `vitest` ReleaseProviderMatrix suites (`ReleaseProviderMatrix.test.tsx` 17 tests, `useReleaseProviderMatrix.test.tsx`, `useReleaseProviderMatrix.toggle.test.tsx`) + guardrail audits (`destructive-confirm-dialog-audit`, `surface-elevation-guardrails`) → **58/58 pass ✓**
- Pre-commit gate (eslint custom rules incl. boundaries, a11y, typecheck) passed.

## Risks

- Low: same-directory presentation-component move, no logic changes, typecheck + full ReleaseProviderMatrix test suite green.
